fix: better order of fields list
