+++ packages/frontend/package.json
@@ -24,7 +24,7 @@
     "cube-core": "^2.13.0",
     "datalib": "^1.9.2",
     "fuse.js": "^3.4.6",
-    "graphic-walker": "^0.0.1",
+    "@kanaries/graphic-walker": "^0.0.1",
     "immer": "^4.0.1",
     "mobx": "^6.3.2",
     "mobx-react-lite": "^3.2.0",
